+++ opal/patients/serializer.py
@@ -1,8 +1,6 @@
 """Module that contains serializer classes for patient models."""
 
 from rest_framework import serializers
-
-from opal.hospital_settings.models import Institution, Site
 from opal.patients.models import HospitalPatient, Patient, Relationship
 
 
@@ -11,7 +9,7 @@ class PatientDetailSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Patient
-        fields = ['first_name', 'last_name', 'date_of_birth', 'sex', 'health_insurance_number']
+        fields = ['first_name', 'last_name', 'date_of_birth', 'sex', 'ramq']
 
 
 class PatientRegistrationSerializer(serializers.ModelSerializer):
@@ -19,23 +17,7 @@ class PatientRegistrationSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Patient
-        fields = ['first_name', 'last_name', 'health_insurance_number']
-
-
-class InstitutionSiteSerializer(serializers.ModelSerializer):
-    """Hospital patient serializer used to get encryption values for registration web site."""
-
-    class Meta:
-        model = Institution
-        fields = ['id', 'name']
-
-
-class SiteSerializer(serializers.ModelSerializer):
-    """Hospital patient serializer used to get encryption values for registration web site."""
-
-    class Meta:
-        model = Site
-        fields = ['id', 'name']
+        fields = ['ramq']
 
 
 class HospitalPatientRegistrationSerializer(serializers.ModelSerializer):