+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Adjust legacy chart api endpoint to use legacy_id (!185)
 * feat: add model for a user's mobile device (!187)
 * Add hospital service for communication with the OIE (!173)
+* Add REST API endpoint for the questionnaire report generation
 
 ### Documentation